refactor(service): split service.rs into submodules
The monolithic service.rs (7263 lines) was unwieldy for review,
navigation, and parallel editing. This commit extracts the logical
sections into focused submodules under service/:

  log_utils.rs       – log file rotation / tail helpers
  single_instance.rs – macOS flock-based wrapper lock
  launch_at_login.rs – macOS Launch at Login, plist helpers, migration
  wrapper.rs         – process wrapper loop, state machine, log events
  purge.rs           – data purge, doctor, process-reaping
  linux.rs           – Linux/systemd service management
  macos.rs           – macOS/launchd service management
  windows.rs         – Windows registry/task service management

service.rs is kept as the module root per project code-style rules
(foo.rs + foo/ directory, never foo/mod.rs). It now contains only the
ServiceCommand enum, top-level constants, platform dispatch shims,
re-exports, and the test module. All logic is behaviour-preserving.
